Updated model

+++ app/src/main/java/hu/kristof/nagy/hikebookclient/model/weather/ListSnow.kt
@@ -4,5 +4,5 @@ import com.squareup.moshi.Json
 
 data class ListSnow(
     @Json(name = "3h")
-    val vol: Int?
+    val vol: Float?
 )
